LayoutTests/imported/w3c/web-platform-tests/fetch/http-cache/no-vary-search.tentative.any.worker-expected.txt


FAIL When params is set to true, URL differs only by their parameters (other than `dispatch` and `uuid`) should not be cached as different entries. assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
PASS Ground truth: When key-order is not set, URLs should be compared in an order-sensitive way.
FAIL When key-order is set , URLs should be compared in an order-insensitive way. Matched cases: assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
PASS When key-order is set , URLs should be compared in an order-insensitive way. Not matched cases
